sanity checks on bricks
git-svn-id: https://sovereigncomputing.net/svn/sneer/trunk@1615 ee3b4d21-14f5-0310-a1fa-c5d8b2925ec4

+++ tests/functional/adapters/SneerCommunity.java
@@ -30,10 +30,12 @@ public class SneerCommunity implements SovereignCommunity {
 		File userHome = SystemUtils.getUserHome();
 		String fileName = ".sneer+"+StringUtils.deleteWhitespace(name);
 		File root = new File(userHome, fileName);
-		try {
-			FileUtils.cleanDirectory(root);
-		} catch (IOException e) {
-			throw new wheel.lang.exceptions.NotImplementedYet(e);
+		if(root.exists()) {
+			try {
+				FileUtils.cleanDirectory(root);
+			} catch (IOException e) {
+				throw new wheel.lang.exceptions.NotImplementedYet(e);
+			}
 		}
 		SneerConfig config = new SneerConfigMock(root);
 		return config;
